+++ src/collective/eeafaceted/collectionwidget/browser/views.py
@@ -1,6 +1,9 @@
 # -*- coding: utf-8 -*-
 from Products.CMFCore.utils import getToolByName
 from Products.Five.browser import BrowserView
+from eea.facetednavigation.browser.app.view import FacetedContainerView
+from eea.facetednavigation.subtypes.interfaces import IFacetedNavigable
+from collective.eeafaceted.collectionwidget.utils import getCollectionLinkCriterion
 
 
 class RenderCategoryView(BrowserView):
@@ -30,4 +33,27 @@ class RenderTermView(BrowserView):
         self.term = term
         self.category = category
         self.widget = widget
+        return self.index()
+
+
+class FacetedDashboardView(FacetedContainerView):
+    """ Facetednavigation view, managing default collection widget redirection """
+
+    def __call__(self):
+        criterion = getCollectionLinkCriterion(self.context)
+        collectionUID = self.context.REQUEST.form.get('{0}[]'.format(criterion.__name__))
+        if collectionUID or self.context.REQUEST.form.get('facetedQuery', '') or not criterion.default:
+            return self.index()
+        if not self.request['HTTP_REFERER'].endswith('configure_faceted.html') and \
+           not self.request['URL'].endswith('folder_contents') and \
+           not self.request.get('no_redirect', '0') == '1':
+            catalog = getToolByName(self.context, 'portal_catalog')
+            brains = catalog(UID=criterion.default)
+            if brains:
+                collection = brains[0].getObject()
+                container = collection.aq_inner.aq_parent
+                if not container == self.context and \
+                   IFacetedNavigable.providedBy(container):
+                    self.request.RESPONSE.redirect(container.absolute_url())
+                    return ''
         return self.index()