+++ src/collective/eeafaceted/collectionwidget/tests/test_widget.py
@@ -218,63 +218,6 @@ class TestWidget(BaseWidgetCase):
         widget()
         self.assertEquals(widget.default, None)
 
-    def test_default_subfaceted(self):
-        """If default is in a subfolder with a faceted defined on it,
-           user will be redirected to this subfolder then the default will be displayed."""
-        data = Criterion(vocabulary=COLLECTION_VOCABULARY)
-        # create a subfolder, add collections into it
-        # and activate the faceted navigation
-        api.content.create(
-            id='subfolder',
-            type='Folder',
-            title='Subfolder',
-            container=self.folder
-        )
-        c3 = api.content.create(
-            id='collection3',
-            type='Collection',
-            title='Collection 3',
-            container=self.folder.subfolder
-        )
-        subtyper = getMultiAdapter((self.folder.subfolder, self.request),
-                                   name=u'faceted_subtyper')
-        subtyper.enable()
-        # enabling a faceted will redirect to it, so, cancel this
-        self.request.RESPONSE.status = 200
-
-        # when default in self.folder, not redirected already at the right place
-        widget = CollectionWidget(self.folder, self.request, data=data)
-        widget.data.default = self.collection1.UID()
-        widget()
-        self.assertEquals(self.request.RESPONSE.status, 200)
-
-        # now set default on a collection of the subfolder
-        widget.data.default = c3.UID()
-
-        # not redirected if we are on folder/folder_contents or folder/configure_faceted.html
-        # or it is no more possible to access the folder actions, always redirected to subfolder...
-        self.request['URL'] = self.folder.absolute_url() + '/folder_contents'
-        # call to 'default' is memoized, so clean it
-        del IAnnotations(self.request)['plone.memoize']
-        widget()
-        self.assertEquals(self.request.RESPONSE.status, 200)
-        folderURL = self.folder.absolute_url()
-        self.request['URL'] = folderURL
-        self.request['HTTP_REFERER'] = self.folder.absolute_url() + '/configure_faceted.html'
-        # call to 'default' is memoized, so clean it
-        del IAnnotations(self.request)['plone.memoize']
-        widget()
-        self.assertEquals(self.request.RESPONSE.status, 200)
-
-        # we are redirected to the subfolder if accessing directly folder
-        self.request['URL'] = folderURL
-        self.request['HTTP_REFERER'] = folderURL
-        # call to 'default' is memoized, so clean it
-        del IAnnotations(self.request)['plone.memoize']
-        widget()
-        self.assertEquals(self.request.RESPONSE.status, 302)
-        self.assertEquals(self.request.RESPONSE.getHeader('location'), self.folder.subfolder.absolute_url())
-
     def test_count(self):
         data = Criterion()
         widget = CollectionWidget(self.folder, self.request, data=data)